(misc) utility to run tasks via shell scripts
