+++ ajc/info_command.go
@@ -1,3 +1,7 @@
+// Copyright (c) 2022, R.I. Pienaar and the Project contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
 package main
 
 import (
